Use div for code block so vue interpolation works
vuepress adds a `v:pre` tag to a fenced code block, which escapes
any vue interpolation so you see the full, unaffected code. this
interferes with instances where you need the vue interpolation for the
code to make sense. I updated an escaped code block to use the same
style as a block earlier in the snippet so that interpolation works.

see https://github.com/vuejs/vuepress/issues/413

Signed-off-by: zachmandeville <webmaster@coolguy.website>