Fix community form submission
Pardot issues an http request to a specified url in their config once
the ajax call is submited, in this case it is
https://kuma.io/community?form_success=true

Before migrating to Jekyll, we just ignored that callback and rendered a
success message, even in the unlikely case that it could fail.
After migrating to Jekyll, we added a success and error
callbacks to handle both scenarios. However, the error handler gets
executed in both scenarios, this is because the response the Ajax gets
is a redirect and a parseError and there is no way to know whether the
call was a success or not.

We are opting here to ignore the error case as before and just show the
success message.

Signed-off-by: Fabian Rodriguez <fabian.rodriguez@konghq.com>