Fix cpinstall tag for kong-mesh
There's a slight difference between Kuma and Kong Mesh in terms of the
base segment of the docs' url.
For Kuma, all the docs live under `/docs/`. Unfortunately, that's not
the case for Kong Mesh. Kong docs renders multiple products, so we
namespace the docs with the corresponding product name, i.e. they live
under `/mesh/`.

This commit updates the tag's code so that it uses the base path in which the
page is generated as the first url segment. That way we can re-use the
code in both platforms.

It also modifies the link that is rendered, so that it uses `page.release`
instead of `page.version` so that it also works for both docs sites.

Signed-off-by: Fabian Rodriguez <fabian.rodriguez@konghq.com>