fix / improve vite config file
styles source maps in dev are very helpful during development

`css.preprocessorOptions.scss.additionalData` needs to be a string or a
function, so having array was wrong and did't work with `devSourcemap`
setting.

modern scss compiler api is just a suggested by sass and legacy api will
be sooner or later removed.

Signed-off-by: Bart Smykla <bartek@smykla.com>