fix(styles): make code more readable on dark background (#2105)
The setup commands on the home page are displayed with a shadow that makes the text blurry and difficult to read. Pipes (|) within the commands also have a bright background, making them unnecessarily prominent. This PR attempts to improve the readability and aesthetics of such text by removing these extra visual elements.

---------

Signed-off-by: Fuchen Shi <fuchen@emq.com>