docs: add logging section for control/data planes & DNS
This commit adds a new section to the troubleshooting guide for logging. It includes information on how to enable debug logging for the control and data planes, as well as how to enable debug logging for DNS.

This partially addresses issue #1302 by providing initial documentation for logging. Further work is needed to complete the full scope of the issue.

Signed-off-by: dascole <joe.dascole@konghq.com>