feat(jekyll/plugin/rbacresources): support configurable and default search paths for input file (#2312)
Add support for locating the input YAML file using a default filename (`rbac.yaml`)
and a set of base paths from `mesh_raw_generated_paths` in `_config.yml`.
The plugin builds the full path using the current `page.release`.
Falls back to the given path if not found. Raises an error if the file
cannot be located.

Signed-off-by: Bart Smykla <bartek@smykla.com>